Throttle GUI send rate.
For a variety of reasons, the GUI is able to send data more quickly
than the Arduino could process it.  As a result, the Arduino's
receive buffer could become full and characters could get dropped
from a long program (usually causing syntax errors).

To prevent this, the GUI now only sends at most 1 line for each ':'
sent by the Arduino (corresponding to a prompt for a new line).  This
makes the GUI behave more like an interactive user at the serial
console.